// src/app/interfaces/clinician.ts
export interface Clinician {
    id: string;
    firstName: string;
    lastName: string;
    city: string;
    state: string;
    clinicianLocation:string;
    registrationNo: string;
    zipCode: string;
    mobileNumber: string;
    emailId: string;
    password: string;
    userType : string;
    dob : string;
    sex : string;
    serviceType:string;
    criticalCare:any;
    specialization:string;
}